requirements.txt
pytest==8.2.2
pytest-playwright==0.5.1
pytest-timeout==2.3.1
